Sanitize basic config before module loading.
I am not completely convinced yet that this is actually a good idea. It adds complexity to the build system in an unconventional way. I would prefer for it to live in the core, but since dependencies of the core technically run first, it is the only way to guarantee that all modules will see the sanitized config, unless we delay initialization or loading of modules on purpose (and otherwise unecessarily). There may be a better way.